refactor names files about insert mode && cargo fmt
