+++ src/base/commands/commands/mod.rs
@@ -1,5 +1,6 @@
 pub mod docs;
 pub mod edit_mode;
+pub mod environment;
 pub mod jumps;
 pub mod request;
 pub mod response;
@@ -7,4 +8,3 @@ pub mod submit;
 pub mod tabs;
 pub mod ui;
 pub mod utils;
-pub mod environment;
