+++ src/base/states/states/default_insert_mode.rs
@@ -1,7 +1,7 @@
 use crate::actions::Actions;
 use crate::base::states::states::{CommandsMap, State, StatesNames};
 use crate::commands::Commands;
-use crate::input::keymaps::input_mode::keymap_factory;
+use crate::input::keymaps::insert_mode::keymap_factory;
 use std::collections::HashMap;
 
 pub struct DefaultEditMode {
