+++ src/view/components/mod.rs
@@ -15,7 +15,7 @@ pub trait StatedComponents<State> {
 pub mod block_text;
 pub mod counter_response_time;
 pub mod doc_reader;
-pub mod input_block;
+pub mod input_insert_mode;
 pub mod tab_blocked_text;
 pub mod tab_list;
 pub mod welcome_doc;
