+++ src/view/views/mod.rs
@@ -1,10 +1,10 @@
 use std::collections::HashMap;
 
 pub mod app;
+pub mod environment;
 pub mod logs;
 pub mod request;
 pub mod response;
 pub mod tabs_request;
-pub mod environment;
 
 pub type ViewStates = HashMap<String, String>;
